+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 ## [[NEXT]](https://github.com/iExecBlockchainComputing/tee-worker-post-compute/releases/tag/vNEXT) 2024
 
-### New Features
+### Quality
 
 - Stop Building Gramine Variant for tee-worker-post-compute in Jenkins Pipeline. (#118)